requirements.txt
pika
telebot
tf-keras
numpy
matplotlib
torch
torchvision
pillow
